結合テーブルを修正。全てのテーブルの外部キーを削除し、単方向の参照とした。テストプログラム未修正。

+++ src/main/java/jp/spring/boot/algolearn/repository/CourseRepository.java
@@ -8,6 +8,6 @@ import jp.spring.boot.algolearn.bean.CourseBean;
  * コース用リポジトリ(course repository)
  * @author tejc999999
  */
-public interface CourseRepository extends JpaRepository<CourseBean, Integer> {
+public interface CourseRepository extends JpaRepository<CourseBean, Long> {
 
 }
