+++ src/main/java/jp/spring/boot/algolearn/repository/QuestionRepository.java
@@ -9,6 +9,6 @@ import jp.spring.boot.algolearn.bean.QuestionBean;
  * @author tejc999999
  */
 public interface QuestionRepository extends
-                                    JpaRepository<QuestionBean, Integer> {
+                                    JpaRepository<QuestionBean, Long> {
 
 }
